requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
pyodbc==5.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
email-validator==2.1.0.post1
click==8.1.7
httpx==0.25.3
bcrypt==4.3.0 